requirements.txt
alembic==1.4.3
appdirs==1.4.4
attrs==20.2.0
autoflake==1.4
bandit==1.6.2
black==20.8b1
Brotli==1.0.9
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cleo==0.7.6
click==7.1.2
clikit==0.4.3
cryptography==3.1
dash==1.16.1
dash-bootstrap-components==0.10.6
dash-core-components==1.12.1
dash-html-components==1.1.1
dash-renderer==1.8.1
dash-table==4.10.1
dnspython==2.0.0
dominate==2.5.2
email-validator==1.1.1
flake8==3.8.3
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==1.9.0
Flask-Compress==1.5.0
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-Testing==0.8.0
Flask-WTF==0.14.3
future==0.18.2
gevent==20.6.2
gitdb==4.0.5
GitPython==3.1.8
greenlet==0.4.16
gunicorn==20.0.4
html5lib==1.1
idna==2.10
isort==5.5.2
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==1.0.0
mypy-extensions==0.4.3
numpy==1.19.2
pandas==1.1.2
pastel==0.2.1
pathspec==0.8.0
pbr==5.5.0
pexpect==4.8.0
pkginfo==1.5.0.1
plotly==4.10.0
poetry==1.0.10
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pylev==1.3.0
pyparsing==2.4.7
pyrsistent==0.14.11
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
requests-toolbelt==0.8.0
retrying==1.3.3
SecretStorage==3.1.2
shellingham==1.3.2
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.19
stevedore==3.2.2
toml==0.10.1
tomlkit==0.5.11
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
visitor==0.1.3
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.1.0
